Moved text editor to components folder
