--- src/customElements/components/app-guide/quick-fact/editor/index.ts
@@ -1,4 +1,0 @@
-export { 
-    AppGuideEditorElement, 
-    AppGuideEditorElement as default 
-} from './element';
