+++ src/customElements/components/markdown-editor/index.ts
@@ -1,0 +1,4 @@
+export { 
+    MarkdownEditorElement, 
+    MarkdownEditorElement as default 
+} from './element';
