+++ src/customElements/components/markdown-editor/keymap.ts
